Extract cqrs_event! to it's own module so it can be re-used
